Add optional 'non-strict' mode
Turning off strict more disables some of the validations,
porentially allowing for data-integrity issues. However
it is hoped that it will allow for performance improvements
and, if suplemented with database constraints should not result
in significant changes in overall data integrity. Instead the main
trade off will be ugly exceptions.